change to mergeTw
