+++ packages/frontend/package.json
@@ -106,6 +106,7 @@
     "sharp": "^0.33.2",
     "stable-hash": "^0.0.3",
     "swr": "^2.1.5",
+    "tailwind-merge": "^2.2.2",
     "tailwind-styled-components": "^2.2.0",
     "tinycolor2": "^1.6.0",
     "use-deep-compare": "^1.1.0",
