+++ packages/frontend/src/features/Navigation/NavigationBarButton.tsx
@@ -4,7 +4,7 @@ import Link from 'next/link';
 import { Icon } from '@iconify/react';
 import { Tooltip } from '@mantine/core';
 import { extractClassName } from './utils';
-import { mergeDefaultValues } from '../../utils/defaultValues';
+import { mergeDefaultTailwindClassnames } from '../../utils/mergeDefaultTailwindClassnames';
 
 /**
  * NavigationBarButton: a button for the navigation bar
@@ -31,7 +31,7 @@ const NavigationBarButton = ({
     arrow: 'bg-base-light',
   };
 
-  const mergedClassnames = mergeDefaultValues(classNamesDefaults, classNames);;
+  const mergedClassnames = mergeDefaultTailwindClassnames(classNamesDefaults, classNames);;
   return (
     <React.Fragment>
       <Tooltip
