+++ packages/sampleCommons/config/brh/footer.json
@@ -1,4 +1,7 @@
 {
+  "classNames": {
+    "root": "bg-base-darker"
+  },
   "footerRightLogos": [
     [
       {
